Use `useThemeToggle` hook in `CommandMenu` and  `ModeToggle` (#274)
### TL;DR

This pull request improves theme toggling by incorporating the `useThemeToggle` hook
- Remove theme toggle logic from `ModeToggle`, import and use `useThemeToggle` hook instead.
- Add `CommandMenuItem` which uses `useThemeToggle` hook to toggle theme from within the `CommandMenu` component.

### What changed?

The code implements to use of the `useThemeToggle` in the `CommandMenu` and replaces the logic in the `ModeToggle` components. A new settings option in the `CommandMenu` has been added to toggle the theme.

### How to test?

Check if the theme toggle is working properly in both the `CommandMenu` and the `ModeToggle` components.

### Why make this change?

This change is made to make the theme toggling more reusable by using a dedicated hook. This change was made to facilitate the use of theme toggle functionality within a `CommandMenuItem`, or any other component without added complexity or inefficient code. This also separates logic from UI in `ModeToggle.tsx`