perf: Remove HTML parsing, instead sanitize and set dangerously
Parsing HTML to React elements is a performance bottleneck which slows down code block rendering. This change removes the parsing and instead sanitizes the HTML with DOMPurify, and sets it with `dangerouslySetInnerHTML`

Also, this HTML processing logic is moved from `CodeHighlight` to `useShiki`, this way the hook always returns sanitized HTML, and logic is centralized.