fix: Revert exporting `astro:content` utils from `utils/index.ts`
Using a barrel file to simplify util exports was causing `astro:content` to be bundled with the client code even though the utils that use it were only being called on the server.

To fix this, any utils that import `astro:content` were removed from the barrel file and instead imported with the full path to the util.

This commit also consolidates non `astro:content` utils into a single import from `@/lib/utils` to simplify imports in `MDLayout` and `CommandMenu`.