fix: Use `setTimeout` instead of `requestAnimationFrame` to unify browser behavior (#545)
`lib/utils/toggleTheme.ts` was using `requestAnimationFrame` to schedule the removal of the transition styles. This worked well in Chrome, but in Firefox, the styles were not applied before the removal. This change uses `setTimeout` to schedule the removal of the styles, which ensures the styles are applied before removal in Firefox.

This commit also removes manual vendor prefixing, this is not needed and has been tested in all major browsers.